backend/requirements.txt
cryptography==43.0.1

environs==9.5.0
loguru==0.7.2
uvicorn==0.30.6
fastapi==0.114.2

alembic==1.13.2
SQLAlchemy==1.4.41
sqlmodel==0.0.8
psycopg2==2.9.1
asyncpg==0.29.0

passlib==1.7.4
PyJWT==2.8.0

pyotp==2.9.0
qrcode==7.4.2

emails==0.6
Jinja2==3.1.4
